refactro!: разделение кода на отдельные модули
